refacto: Move images in a folder
