+++ src/assets/images/crossbow_left.png
