+++ src/assets/images/crossbow_up.png
